requirements.txt
streamlit==1.32.0
weaviate-client==4.4.0
openai==1.12.0
pandas==2.1.0
python-dotenv==1.0.0
matplotlib==3.8.0
networkx==3.2.0
